requirements.txt
pywin32
python-dotenv
Flask